upgrade error

+++ db/upgrade.php
@@ -919,7 +919,7 @@ function xmldb_block_exaport_upgrade($oldversion) {
         upgrade_block_savepoint(true, 2019031500, 'exaport');
     }
 
-    if ($oldversion < 2019101005) {
+    if ($oldversion < 2019111202) {
         // change indexes again
         $tableswithindexes = array(
                 'block_exaportcate' => array('pid', 'userid', 'shareall', 'internshare', 'structure_shareall', 'structure_share'),
@@ -951,7 +951,9 @@ function xmldb_block_exaport_upgrade($oldversion) {
             foreach ($existingindexes as $indexname => $eindex) {
                 if (trim(strtolower($indexname)) != 'primary') {
                     $index = new xmldb_index($indexname, XMLDB_INDEX_NOTUNIQUE, $eindex['columns']);
-                    $dbman->drop_index($table, $index);
+                    if ($dbman->index_exists($table, $index)) {
+                        $dbman->drop_index($table, $index);
+                    }
                     //$DB->execute('DROP INDEX '.$eindex.' ON '.$DB->get_prefix().$tablename.' ');
                 }
             }
@@ -962,7 +964,7 @@ function xmldb_block_exaport_upgrade($oldversion) {
                 }
             }
         }
-        upgrade_block_savepoint(true, 2019101005, 'exaport');
+        upgrade_block_savepoint(true, 2019111202, 'exaport');
     }
 
     // TODO: delete structure fields / tables.
